Use 1 grep process for the entire runtime, instead of 1 per create event
This change improves performance by not doing a grep call per create event.
Also, instead of matching the file from the event onto the list of
watchfiles (which was buggy, ie: a create event for foo will cause foo to be
deleted if you have foobar as a watchfile), we now use the watchfiles file as
file of patterns to match onto the file from the event.
As an added bonus, this allows you to use regular expressions in your
watch patterns.